Fix compilation warnings (#13242)
- Avoid incorrect format specifiers in ParModelica and OMDumpXML.
- Add braces to avoid dangling else in OMEdit Debugger.
- Use C++ `for` instead of `Qt` foreach when traversing `QJsonArray` to
  get rid of deprecation warnings.
- Use static `QFontDatabase::families` method with Qt 6.0.
- Add missing `override`.
- Don't ignore return value of `QTranslator::load` in OMShell.

+++ OMEdit/OMEditLIB/TransformationalDebugger/OMDumpXML.cpp
@@ -336,7 +336,7 @@ MyHandler::MyHandler(QFile &file, QHash<QString,OMVariable> &vars, QList<OMEquat
         currentEquation->eqs = nestedEquations;
         operations.clear();
         if (currentEquation->index != equations.size()) {
-          printf("failing: %d expect %d\n", currentEquation->index, equations.size()+1);
+          printf("failing: %d expect %zu\n", currentEquation->index, static_cast<size_t>(equations.size())+1);
           break;
         }
         equations.append(currentEquation);
